pyproject.toml
[project]
name = "navigator"
version = "0.1.0"
description = ""
authors = [
    {name = "Juan Ortega",email = "juan.ortega-contreras@capgemini.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk (>=1.5.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]

[tool.poetry]
packages = [{include = "navigator", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
